Extract SubmitButton component out of CheckoutComponent
I think this'll be simpler to understand as we add a new CTA for Stripe
Hosted Checkout.
